requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-ipware==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-simple-apikey==1.0.2
idna==3.4
packaging==23.2
pycparser==2.21
python-dotenv==1.0.0
python-ipware==2.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
